Implement Gameplay and Camera

+++ app/src/main/java/com/example/mmue_lm3/gameobjects/BoosterItemObject.java
@@ -48,7 +48,7 @@ public class BoosterItemObject extends ItemObject {
         paint.setStyle(Paint.Style.FILL);
 
         Rect rect = this.getRectangle();
-        rect.offset(-camera.getX(), -camera.getY());
+        rect.offset(-(int)camera.getX(), -(int)camera.getY());
         canvas.drawRect(rect, paint);
     }
 
@@ -58,7 +58,7 @@ public class BoosterItemObject extends ItemObject {
         paint.setStyle(Paint.Style.FILL);
 
         Rect rect = this.getRectangle();
-        rect.offset(-camera.getX(), -camera.getY());
+        rect.offset(-(int)camera.getX(), -(int)camera.getY());
         canvas.drawRect(rect, paint);
     }
 
@@ -68,7 +68,7 @@ public class BoosterItemObject extends ItemObject {
         paint.setStyle(Paint.Style.FILL);
 
         Rect rect = this.getRectangle();
-        rect.offset(-camera.getX(), -camera.getY());
+        rect.offset(-(int)camera.getX(), -(int)camera.getY());
         canvas.drawRect(rect, paint);
     }
 
@@ -78,7 +78,7 @@ public class BoosterItemObject extends ItemObject {
         paint.setStyle(Paint.Style.FILL);
 
         Rect rect = this.getRectangle();
-        rect.offset(-camera.getX(), -camera.getY());
+        rect.offset(-(int)camera.getX(), -(int)camera.getY());
         canvas.drawRect(rect, paint);
     }
