+++ app/src/main/java/com/example/mmue_lm3/gameobjects/DestroyablePlatformObject.java
@@ -29,7 +29,7 @@ public class DestroyablePlatformObject extends PlatformObject {
         paint.setStyle(Paint.Style.FILL);
 
         Rect rect = this.getRectangle();
-        rect.offset(-camera.getX(), -camera.getY());
+        rect.offset(-(int)camera.getX(), -(int)camera.getY());
         canvas.drawRect(rect, paint);
     }
