+++ app/src/main/java/com/example/mmue_lm3/gameobjects/EctsItemObject.java
@@ -33,7 +33,7 @@ public class EctsItemObject extends ItemObject {
         paint.setStyle(Paint.Style.FILL);
 
         Rect rect = this.getRectangle();
-        rect.offset(-camera.getX(), -camera.getY());
+        rect.offset(-(int)camera.getX(), -(int)camera.getY());
         canvas.drawRect(rect, paint);
     }
